hcs/requirements.txt
absl-py==1.0.0
apptools==5.1.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
Bottleneck==1.3.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.5
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance@291c8d338c2dca3a6cbe2df1b891f9fdcc8a411b
chamferdist==1.0.0
charset-normalizer==2.0.7
chumpy==0.70
click==7.1.2
colorama==0.4.4
configobj==5.0.6
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
envisage==6.0.0
flatbuffers==2.0
freetype-py==2.2.0
gast==0.4.0
gevent==21.1.2
glfw==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.2
importlib-resources==5.1.2
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
# Editable install with no version control (jax==0.2.21)
-e c:\users\han\share\jax-jaxlib-v0.1.72
jaxlib @ file:///C:/Users/han/Share/jax/jaxlib-0.1.73-cp37-none-win_amd64.whl
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==12.0.0
llvmlite==0.34.0
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mayavi==4.7.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1618854593605/work
mkl-service==2.3.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
ninja==1.10.0.post2
nitime==0.9
nose==1.3.7
notebook==6.3.0
numba==0.51.2
numpy @ file:///C:/ci/numpy_and_numpy_base_1620831380353/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
open3d==0.9.0.0
open3d-python==0.7.0.0
opencv-python==4.5.1.48
opendr==0.73
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1617386319301/work
plyfile==0.7.4
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyface @ file:///C:/ci/pyface_1614366094446/work
pyglet==1.5.16
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyQt5Designer==5.14.1
pyrealsense2==2.45.0.3217
pyrender==0.1.45
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==1.0.1
PyYAML==5.4.1
pyzmq==22.0.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy @ file:///C:/ci/scipy_1618856134946/work
Send2Trash==1.5.0
serial==0.0.97
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smplx==0.1.27
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
traits @ file:///C:/ci/traits_1611259464695/work
traitsui==7.2.1
transforms3d==0.3.1
trimesh==3.9.18
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3==1.26.7
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-server==0.4
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
win32-setctime==1.0.3
wincertstore==0.2
wrapt==1.13.3
yacs==0.1.8
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
